Differential Revision: [D18339142](https://our.internmc.facebook.com/intern/diff/D18339142)

+++ benchmarks/operator_benchmark/pt/qpool_test.py
@@ -31,7 +31,7 @@ qmaxpool2d_configs = op_bench.config_list(
         'contig': (False, True),
         'dtype': (torch.qint32, torch.qint8, torch.quint8),
     },
-    tags=('short',)
+    tags=('long',)
 )
